webapi/requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
dictdiffer==0.9.0
fastapi==0.87.0
h11==0.14.0
idna==3.4
install==1.3.5
iso8601==1.1.0
pydantic==1.10.2
pypika-tortoise==0.1.6
pytz==2022.6
sniffio==1.3.0
starlette==0.21.0
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.20.0
